+++ src/index/event.rs
@@ -8,14 +8,14 @@ pub enum Event {
     inscription_id: InscriptionId,
     location: Option<SatPoint>,
     parent_inscription_ids: Vec<InscriptionId>,
-    sequence_number: u32,
+    sequence_number: u64,
   },
   InscriptionTransferred {
     block_height: u32,
     inscription_id: InscriptionId,
     new_location: SatPoint,
     old_location: SatPoint,
-    sequence_number: u32,
+    sequence_number: u64,
   },
   RuneBurned {
     amount: u128,
